Extract main function into cmd/notes
